Move IdeTheme class into a shared file instead of duplicating
