+++ packages/devtools_app/lib/src/config_specific/ide_theme/ide_theme.dart
@@ -2,6 +2,17 @@
 // Use of this source code is governed by a BSD-style license that can be found
 // in the LICENSE file.
 
+import 'package:flutter/widgets.dart';
+
 export 'ide_theme_stub.dart'
     if (dart.library.html) 'ide_theme_web.dart'
     if (dart.library.io) 'ide_theme_desktop.dart';
+
+/// IDE-supplied theming.
+class IdeTheme {
+  IdeTheme({this.backgroundColor, this.foregroundColor, this.fontSize});
+
+  Color backgroundColor;
+  Color foregroundColor;
+  double fontSize;
+}
