+++ packages/devtools_app/lib/src/config_specific/ide_theme/ide_theme_desktop.dart
@@ -2,15 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be found
 // in the LICENSE file.
 
-import 'package:flutter/widgets.dart';
+import 'ide_theme.dart';
 
-/// IDE-supplied theming.
-class IdeTheme {
-  IdeTheme({this.backgroundColor, this.foregroundColor, this.fontSize});
-
-  factory IdeTheme.load() => IdeTheme();
-
-  Color backgroundColor;
-  Color foregroundColor;
-  double fontSize;
-}
+/// Load any IDE-supplied theming.
+IdeTheme getIdeTheme() => null;
